refactor : 여러 개 쿼리 -> 조인 사용

+++ src/main/java/com/example/cns/chat/domain/repository/ChatRoomListRepository.java
@@ -41,7 +41,8 @@ public class ChatRoomListRepository {
                 .where(chatParticipation.member.eq(memberId))
                 .orderBy(chatRoom.lastChatId.desc())
                 .limit(pageSize)
-                .offset((offset - 1) * pageSize)
+                .offset((long) (offset - 1) * pageSize)
                 .fetch();
     }
+
 }
